Add client structure and refactor movement functions for clarity
