+++ server/include/client_info.h
