+++ server/include/ressources.h
@@ -7,9 +7,7 @@
 
 #ifndef RESSOURCES_H_
     #define RESSOURCES_H_
-
-#include "flag.h"
-
+    #include "flag.h"
 
 typedef struct tile_s {
     int food;
@@ -27,7 +25,5 @@ void generate_map_resources(server_config_t *conf);
 void send_whole_map(int fd, server_config_t *conf);
 void generate_tile_resources(tile_t *tile);
 void send_tile_content(int fd, int x, int y, tile_t *tile);
-
-
 
 #endif /* !RESSOURCES_H_ */
