+++ server/src/functional/handle_client_data.c
@@ -101,18 +101,29 @@ static bool handle_auth(auth_context_t *ctx, char *buffer)
     return true;
 }
 
-bool handle_client_data(client_t **clients, int fd, server_config_t *conf)
+static bool process_client_request(client_t *client, int fd,
+    server_config_t *conf, char *buffer)
+{
+    auth_context_t ctx = {&conf->clients, client, conf};
+
+    if (client->is_graphic == false)
+        respond_to_server_fd(fd, conf, buffer, client);
+    if (client->state == WAITING_NAME)
+        handle_auth(&ctx, buffer);
+    return false;
+}
+
+bool handle_client_data(client_t **clients, int fd,
+    server_config_t *conf)
 {
     char buffer[1024];
     ssize_t r;
     client_t *client = *clients;
-    auth_context_t ctx = {clients, NULL, conf};
 
     while (client && client->fd != fd)
         client = client->next;
     if (!client)
         return true;
-    ctx.client = client;
     r = read(fd, buffer, sizeof(buffer) - 1);
     if (r <= 0) {
         remove_client(clients, fd);
@@ -120,9 +131,5 @@ bool handle_client_data(client_t **clients, int fd, server_config_t *conf)
     }
     buffer[r] = '\0';
     printf("[DEBUG] Data from fd %d: %s\n", fd, buffer);
-    if (client->is_graphic == false)
-        respond_to_server_fd(fd, conf, buffer, client);
-    if (client->state == WAITING_NAME)
-        handle_auth(&ctx, buffer);
-    return false;
+    return process_client_request(client, fd, conf, buffer);
 }
